+++ firmware/controllers/trigger/decoders/trigger_honda.h
@@ -9,8 +9,6 @@
 
 class TriggerWaveform;
 
-void configureOnePlus16(TriggerWaveform *s);
-
 void configureHondaCbr600(TriggerWaveform *s);
 
 void configureHondaK_12_1(TriggerWaveform *s);