+++ firmware/controllers/trigger/decoders/trigger_structure.cpp
@@ -680,10 +680,8 @@ void TriggerWaveform::initializeTriggerWaveform(operation_mode_e triggerOperatio
 	case trigger_type_e::TT_MITSU_4G9x_CAM:
 	    initializeMitsubishi4g9xCam(this);
         break;
-	case trigger_type_e::TT_1_16:
-		configureOnePlus16(this);
-		break;
 
+	case trigger_type_e::TT_UNUSED_50:
 	case trigger_type_e::TT_HONDA_CBR_600:
 		configureHondaCbr600(this);
 		break;